bitrix-sync/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
SQLAlchemy>=2.0.23
PyMySQL>=1.1.0
Flask-SQLAlchemy>=3.1.1
